Move the declarations of is_arma_supported
